Update Validator only header
