+++ core/service/CMakeLists.txt
@@ -3,8 +3,9 @@
 ADD_LIBRARY(validator STATIC
     validator.cpp
 )
+
 target_link_libraries(validator
     hash
-    flatbuffers
+    flatbuffer_service
     logger
 )
